infra/pyproject.toml
[project]
name = "infra"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-aws-lambda-python-alpha>=2.248.0a0",
    "aws-cdk-lib>=2.244.0,<3.0.0",
    "constructs>=10.5.0,<11.0.0",
]
